Combine heat-api and heat-api-cfn rocks
Make single rock for heat-api and heat-api-cfn.
In the process of making single charm for heat-api
and heat-api-cfn, the metadata yaml for the charm
should hold only heat-api container and can act as
heat-api service or heat-api-cfn service based on
configuration parameter. However the container to
resource mapping is static in metadata.yaml. So
the heat-api-image resource should hold packages
for both heat-api and heat-api-cfn.

Removed heat-api-cfn rock and added heat-api-cfn
package in heat-api rock.

Ignore github workflow action to build rock if
the rock is deleted.